agent_16_generate_final_report/requirements.txt
pandas>=1.5.0
matplotlib>=3.6.0
numpy>=1.21.0
openai>=1.0.0
python-dotenv>=0.19.0
seaborn>=0.11.0
